Psudocode to demonstrate PID in lock file
CA doesn't currently store the creating process ID in lock files.  This affects
CA's ability to determine if a lock file is safe to remove as well as the
administrators ability to track down stuck proceses.

This (untested and unvalidated) change gives an idea of how CA could be changed
to add process IDs to its lock files and use that information in determining if
said files should be removed.