Remove date from log files
Having CA insert the date in the middle of its log file names makes log
management more complex:
- logrotate can't reliably match filenames any more
- using shell matching in cron can cause issues on month boundaries
- deleting old logs with find can result in log file removals (on
  quiet/dev systems)

An alternative solution is having CA perform its own log rotation, but
the change required to make that happen is substantially more involved.

We will be trying to run without the date and see if everything
continues to operate as expected.